pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "summitly-backend"
version = "3.0.0"
description = "AI-powered Real Estate Chatbot with Summitly and Repliers Integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Development Team", email = "dev@summitly.local"}
]
keywords = [
    "real-estate",
    "chatbot",
    "ai",
    "voice-assistant",
    "property-search"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business",
]

dependencies = [
    "Flask>=2.3.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "pandas>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

repliers = [
    "exa_py>=1.0.0",
]

audio = [
    "speechrecognition>=3.10.0",
    "pyttsx3>=2.90",
    "pydub>=0.25.1",
]

huggingface = [
    "transformers>=4.36.0",
    "huggingface-hub>=0.19.4",
]

fastapi = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
]

email = [
    "Flask-Mail>=0.9.1",
]

excel = [
    "openpyxl>=3.1.0",
]

all = [
    "summitly-backend[dev,repliers,audio,huggingface,fastapi,email,excel]",
]

[project.urls]
Homepage = "https://summitly.ca"
Documentation = "https://github.com/summitly/backend"
Repository = "https://github.com/summitly/backend.git"
Issues = "https://github.com/summitly/backend/issues"

[tool.setuptools]
packages = ["app", "services", "endpoints", "models", "utils"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=app --cov=services --cov=endpoints --cov=models --cov=utils"
